Report a hit for PhysicsSystem::castRay
This is used by AiTravel to test whether the actor can walk between two
otherwise disconnected pathgrid nodes. Returning false meant there was
no blockage and the path was clear, causing the actor to continually try
to reach pathgrid nodes they couldn't.

It's also used to test whether the there's an unblocked view of the sun,
for the sun-glare effect, when occlusion queries aren't available.